feat: isolate properly computeUsersDistribution
